requirements.txt
request==2.31.0
beautifulsoup4==4.12.2
jinja2==3.1.4
